unit test added generate state
