+++ src/leonardo/configuration.spec.ts
@@ -1,12 +1,13 @@
 /// <reference path="../../typings/angularjs/angular.d.ts" />
 /// <reference path="../../typings/jasmine/jasmine.d.ts" />
-import { leoConfiguration } from '../../src/leonardo/configuration.srv'
+import {leoConfiguration} from '../../src/leonardo/configuration.srv'
+import {IState} from './configuration.srv';
 
-const genericState = {
-  "name": "Create Character",
-  "url": "/character",
-  "verb": "POST",
-  "options": [
+const generateState = ({
+  name = "Create Character",
+  url =  "/character",
+  verb = "POST",
+  options = [
     {
       "name": "success",
       "status": 200,
@@ -22,13 +23,47 @@ const genericState = {
       "delay": "2000"
     }
   ]
-};
+}:IState) :IState => {
 
-describe('addStates', function() {
+  return {
+    name,
+    url,
+    verb,
+    options
+  }
+}
+
+describe('configurations', function() {
   it('should add a state', function() {
     var leo = leoConfiguration();
 
-    leo.addStates([genericState]);
+    leo.addState(generateState({} as IState), false);
+
+    var states = leo.getStates();
+    expect(states.length).toBe(1);
+  });
+
+  it('should add two different states', function() {
+    var leo = leoConfiguration();
+
+    leo.addStates([
+      generateState({name: "logout"} as IState),
+      generateState({} as IState)
+    ]);
+
+    var states = leo.getStates();
+    expect(states.length).toBe(2);
+  });
+
+  it('should merge two similar states', function() {
+    var leo = leoConfiguration();
+
+    const state = generateState({name: "logout"} as IState);
+
+    leo.addStates([
+      state,
+      state
+    ]);
 
     var states = leo.getStates();
     expect(states.length).toBe(1);
@@ -37,8 +72,10 @@ describe('addStates', function() {
   it('should remove a state', function() {
     var leo = leoConfiguration();
 
-    leo.addStates([genericState]);
-    leo.removeState(genericState);
+    const state = generateState({name: "logout"} as IState);
+
+    leo.addState(state, false);
+    leo.removeState(state);
 
     var states = leo.getStates();
     expect(states.length).toBe(0);
